examples/rag-poc/requirements.txt
haystack-ai
sentence-transformers
pypdf
python-dotenv